Fix more compilation warnings

+++ helper.cpp
@@ -23,8 +23,8 @@
  * @param uint64_t n
  * @return vector<int>
  */
-vector<int> primeFactors( uint64_t n ) {
-	vector<int> result;
+vector<uint64_t> primeFactors( uint64_t n ) {
+	vector<uint64_t> result;
 
 	// Find the number of 2s that divide n
 	while( n % 2 == 0 ) {
@@ -84,7 +84,7 @@ vector<uint64_t> sieve( uint64_t n ) {
  * @param uint64_t n The number to perform the primality test on.
  * @return bool Whether or not the number is prime.
  */
-bool isPrime( long long n ) {
+bool isPrime( uint64_t n ) {
 	if( n <= 1 ) {
 		return false;
 	} else if( n <= 3 ) {
